requirements.txt
pandas
python-dotenv
psycopg2-binary
psycopg2